requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.

#### ESSENTIAL LIBRARIES FOR MAIN FUNCTIONALITY ####

# Used when dealing with the affinity matrices.
numpy

# Used when dealing with networks
python-igraph

# Tutorial notebooks
jupyter

